data/datatest: move Equal() to higher-level
