+++ processor/deltatocumulativeprocessor/internal/data/expo/zero_test.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 
+	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/data/datatest"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/data/expo"
 	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/deltatocumulativeprocessor/internal/data/expo/expotest"
 )
@@ -56,7 +57,7 @@ func TestWidenZero(t *testing.T) {
 			}
 			expo.WidenZero(hist, zt)
 
-			is := expotest.Is(t)
+			is := datatest.New(t)
 			is.Equal(want, hist)
 		})
 	}
@@ -108,7 +109,7 @@ func TestSlice(t *testing.T) {
 
 			expo.Abs(bins).Slice(from, to)
 
-			is := expotest.Is(t)
+			is := datatest.New(t)
 			is.Equal(want, bins)
 		})
 	}
